Eliminate 'rR' block type hack
In the offline editor, reporters must be treated as requesters. This had
been implemented by converting `'r'` blocks into `'rR'` blocks, but this
ended up "leaking" into SB2s saved from the offline editor, which in
turn caused problems with the online editor.

Now, the `Block` constructor calls the new `shouldForceAsync()` method
on `ExtensionManager` to determine whether or not to set `forceAsync`
(renamed from `forceRequester`). The block type/spec is never altered.

In support of this change, the logic for showing the watcher checkbox
was updated to base its decision on the constructed block rather than
the block spec.